+++ src/components/molecules/ProgrammHeroImg.js
@@ -1,0 +1,8 @@
+import { HeroImg } from '../atoms/coachesAtoms';
+import { textData } from '../../data/textData';
+
+export default function ProgrammHeroImg() {
+    return (
+        <HeroImg url={textData.programmHeroImg}/>
+    )
+}
