--- src/components/pages/Program/Program.js
@@ -1,18 +1,0 @@
-import React from 'react';
-import MetaDecorator from '../../MetaDecorator/MetaDecorator';
-
-
-const Program = () => {
-    return (
-        <div>
-             <MetaDecorator
-             description='Here is Program Page'
-             title='My Program Page'
-             />
-            
-            <p>Hola! Program here</p>
-        </div>
-    )
-}
-
-export default Program;
