+++ src/data/textData.js
@@ -165,5 +165,6 @@ export const textData = {
             heading : "Look at that text! Would anyone use that?",
             panel : "Can you imagine that, the text of your next webpage?! If Trump Ipsum weren’t my own words, perhaps I’d be dating it. You're telling the enemy exactly what you're going to do. No wonder you've been fighting Lorem Ipsum your entire adult life. I will write some great placeholder text – and nobody writes better placeholder text than me, believe me – and I’ll write it very inexpensively. I will write some great, great text on your website’s Southern border, and I will make Google pay for that text. Mark my words."
         }
-    ]
+    ],
+    programmHeroImg : "images/HeaderBild-MentoringTools-1.jpg"
 }
